refactor: add domain/repository.rs
