+++ src/domain/mod.rs
@@ -1,2 +1,3 @@
 pub mod model;
+pub mod repository;
 pub mod service;
